requirements.txt
faust-streaming==0.11.3
aiokafka>=0.8.0,<0.11.0